DSSAT Modeling steps
The file provides an overview of the DSSAT model and the steps implemented to simulate maize yield in two growing seasons, 2019 and 2021, in Busia County, Kenya. The file also includes the description and sources of the datasets used in the project and the final output files for maize-grown areas in the two seasons.